requirements.txt
httpx
pycryptodomex